pyproject.toml
[project]
name = "lambda-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Casey Sayre",email = "casey-sayre@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "aws-cdk-lib (>=2.178.2,<3.0.0)",
    "constructs (>=10.4.2,<11.0.0)"
]

[tool.poetry]
packages = [{include = "lambda_demo", from = "src"}]

[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"

[tool.poetry.group.customers_lambda_function.dependencies]
fastapi = ">=0.115.8,<0.116.0"
mangum = ">=0.19.0,<0.20.0"
pydantic = "^2.10.6"

[tool.poetry.group.products_lambda_function.dependencies]
fastapi = ">=0.115.8,<0.116.0"
mangum = ">=0.19.0,<0.20.0"
pydantic = "^2.10.6"

[tool.poetry.group.orders_lambda_function.dependencies]
fastapi = ">=0.115.8,<0.116.0"
mangum = ">=0.19.0,<0.20.0"
pydantic = "^2.10.6"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
httpx = "^0.28.1"
pytest-cov = "^6.0.0"


[tool.poetry.group.cognito_setup.dependencies]
boto3 = "^1.37.4"
dotenv = "^0.9.9"

[tool.pytest.ini_options]
testpaths = [
    "tests"
]
pythonpath = [
    "."
]
addopts = """   
    --cov=src.lambda_demo.customers_lambda_function
    --cov=src.lambda_demo.orders_lambda_function
    --cov=src.lambda_demo.products_lambda_function
    --cov-report term-missing
""" 

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
